Split of actions for unsubmitted & submitted applications
We currently don't correctly check if the page in the
`ApplicationFormsController` should be visible to the current user.
This extracts the actions that should only be visible for submitted &
unsubmitted
applications to a separate controller, so we don't need to use
`before_action` per action anymore.

No functional changes intended in this commit.
