+++ app/views/candidate_interface/submitted_application_form/review_submitted.html.erb
@@ -21,4 +21,4 @@
   </div>
 </dl>
 
-<%= render 'review', application_form: @application_form, editable: false %>
+<%= render 'candidate_interface/application_form/review', application_form: @application_form, editable: false %>
