+++ app/views/candidate_interface/submitted_application_form/submit_success.html.erb
